% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot}
\alias{gseaplot}
\title{Plot GSEA graph}
\usage{
gseaplot(expr, list, gset = NULL, nompval = 0.05, fdr = 0.25)
}
\arguments{
\item{expr}{Matrix or dataframe of expression data (rows are genes and columns are samples)}

\item{list}{Result of simple_gsea function}

\item{gset}{You can choose one gene set as input or Null which means those satisfied gene sets will be plot}

\item{nompval}{Significance threshold for nominal p-vals for gene sets (default: 0.05)}

\item{fdr}{Significance threshold for FDR q-vals for gene sets (default: 0.25)}
}
\value{
GSEA graph
}
\description{
This GSEA graph is nearly the same with GSEA official plot with very minimal change to the source code.
}
\examples{
gseaplot <- function(expr = data, list = res, gset = NULL, nompval = 0.05, fdr = 0.25)
gseaplot <- function(expr = data, list = res, gset = "chr10q11")

}
