% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleGSEA.R
\name{simple_gsea}
\alias{simple_gsea}
\title{Do a simplified version of GSEA with most default parameters}
\usage{
simple_gsea(dataexpr, group, geneset, gsminsize = 15, gsmaxsize = 500,
  nperm = 1000)
}
\arguments{
\item{dataexpr}{Matrix or dataframe of expression data (rows are genes and columns are samples)}

\item{group}{A vector which distinguishs some samples to others(Just support two phenotypes)}

\item{geneset}{Gene set databases in Molecular Signatures Database (MSigDB)}

\item{gsminsize}{Minimum size (in genes) for database gene sets to be considered (default: 15)}

\item{gsmaxsize}{Maximum size (in genes) for database gene sets to be considered (default: 500)}

\item{nperm}{Number of random permutations (default: 1000)}
}
\value{
gseaResult object
}
\description{
`simple_gsea` is used to output result of GSEA. You just need to import expression matrix and gene set as element files.
Then you have to provide the group info of samples. See this examples for more information
}
\examples{
data <- read.table(file = "expression.txt", sep = "\\t", header = T, row.names = 1, stringsAsFactors = F, quote = "")
group <- c(rep("m", 15), rep("f", 17))
class <- factor(group)
gene_set <- readLines("C1.gmt")
res <- simple_gsea(dataexpr = data, group = class, geneset = gene_set, gsminsize = 15, gsmaxsize = 500, nperm = 1000)

}
